USB HID refactoring
